Use size_t for len in multiple places
digest module
(librdf_digest_to_string): use size_t for mdlen and index.

hash module
(librdf_hash_from_string): size_t for len and index var.

(macro ONE_AT_A_TIME_HASH): use size_t for len

parser module
(librdf_parser_guess_name2, librdf_parser_raptor_get_next_statement,
librdf_parser_raptor_parse_uri_as_stream_write_bytes_handler):
size_t for len

+++ src/rdf_digest.c
@@ -377,9 +377,9 @@ char*
 librdf_digest_to_string(librdf_digest* digest)
 {
   unsigned char* data = digest->digest;
-  int mdlen = digest->factory->digest_length;
+  size_t mdlen = digest->factory->digest_length;
   char* b;
-  int i;
+  size_t i;
 
   b = (char *)librdf_alloc_memory(1 + (mdlen<<1));
   if(!b) {
@@ -387,7 +387,7 @@ librdf_digest_to_string(librdf_digest* digest)
     return NULL;
   }
   
-  for(i = 0; i<mdlen; i++)
+  for(i = 0; i < mdlen; i++)
     sprintf(b+(i<<1), "%02x", (unsigned int)data[i]);
 
   b[i<<1] = '\0';
